Eliminate cloudflareWorker
